requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.15
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
gunicorn==20.1.0
oauthlib==3.2.1
phonenumberslite==8.12.57
psycopg2==2.9.3
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23
